bip-taro-vm: spell out the two types of validation
In this commit, we fix an ambiguity in the spec w.r.t how state
transitions are valid. There're two types of state transitions:
 1. Transitions that include a prev witness
 2. Transitions that don't include one, and instead specify the
    split_commitment_proof

We only need to do the input+output mapping and VM validation for the
first case. For the second case, a parent state transition exists, which
is what create the split in the first place. In this case we just need
to verify the split commitment proof of the child, then verify the state
transition of the parent itself.

When the child is spent (the above is carried out when people are
validating the child's existence), the rules are applied as normal.