bip-taro: modify asset ID generation, require unique commitments
This commit contains a few important changes:

* The asset ID now includes the output index, which ensures they're
  unique and also allows for created assets to live in distinct outputs.
* A concrete requirement (handwavy statement was there before) that only
  a single taro commitment exists in a script tree at any given time. To
  achieve this, we take advantage of the BIP 340 tagged hash contract,
  to ensure that the commitment is either the only element, or is the
  leftmost/rightmost element.
* Building off the above, we then detail a deterministic way to burn
  assets, as well as prove to others that assets have been burnt. The
  asset burning will play a role in the upcoming re-genesis protocol
  feature.

One additional change during transfers is for parties to reveal the
sibling hash of the taro root commitment so uniqueness verification can
be done on that level. In addition, a similar reveal occurs for the
other outputs (if they exist) that don't explicitly commit (as far as
one party is aware) to any other assets.