bip-taro-vm: unroll verification loop for each input, further specify
In this commit, we add some missing details to the input and output
mapping for normal and collectible asset transfers. We also "unroll" the
verification loop for each referenced previous input. We use the input
index (based on the ordering of the created SMT tree) as the input index
for validation, and similarly use the prior `asset_script_key` as the
previous output script.

This mapping helps us get around some quirks on the Taproot level
related to the inclusion of the input amounts for all inputs into the
final sighash. By maintaining this 1-in-1-out mapping each signature
only needs to be aware of the total amount transferred, and not also the
individual distribution of the referenced inputs.