bip-taro: remove asset meta from asset TLV, replace with meta_hash
In this commit, we modify the way the asset meta functions within the
protocol. Rather than have the asset meta live within _each_ instance of
the asset within the context of a series of proofs. We instead commit to
the asset meta itself in the form of a `meta_hash`. This meta hash can
then be optionally revealed within the proof for the given genesis
asset, we call this the `meta_reveal`.

The `meta_reveal` itself is another nested TLV instance. Verifiers of an
asset genesis proof that verify that the TLV encoding of the meta reveal
matches the `meta_hash` within the genesis asset.

The `meta_reveal` has two currently defined fields: the type, and the
data itself. Today the type is just `0`, which just means it's an opaque
blob. In the future, further types can be added that inform clients or
UIs how to parse, render, and interact with the meta.